requirements.txt
pandas
python-pptx
pypinyin
